refactor: reorganize repo directories
