+++ .gitignore
@@ -10,7 +10,7 @@ pids
 *.pid
 *.seed
 *.pid.lock
-bench_sha256_x15/.output
+.output
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
